+++ foursight_core/react/api/auth0_config.py
@@ -1,6 +1,7 @@
 import logging
 import os
 import requests
+from dcicutils.function_cache_decorator import function_cache
 from dcicutils.misc_utils import get_error_message
 from .misc_utils import get_request_domain, is_running_locally
 
@@ -62,8 +63,6 @@ class Auth0Config:
             raise ValueError("Portal URL required for Auth0Config usage.")
         self._portal_url = portal_url
         self._config_url = f"{portal_url}{'/' if not portal_url.endswith('/') else ''}auth0_config?format=json"
-        self._config_data = {}
-        self._config_raw_data = {}
 
     def get_portal_url(self) -> str:
         return self._portal_url
@@ -71,21 +70,11 @@ class Auth0Config:
     def get_config_url(self) -> str:
         return self._config_url
 
+    @function_cache(nocache={})
     def get_config_data(self) -> dict:
         """
         Returns relevant info (dictionary) from the Auth0 config URL in canonical form.
         It contains these properties: domain, client, sso, scope, prompt, connections.
-        This caches its data.
-        """
-        if not self._config_data:
-            self._config_data = self._get_config_data_nocache() or {}
-        return self._config_data
-
-    def _get_config_data_nocache(self) -> dict:
-        """
-        Returns relevant info (dictionary) from the Auth0 config URL in canonical form.
-        It contains these properties: domain, client, sso, scope, prompt, connections.
-        This does NOT cache its data.
         """
         config_raw_data = self.get_config_raw_data()
         domain = config_raw_data.get("auth0Domain") if config_raw_data else None
@@ -119,19 +108,10 @@ class Auth0Config:
             "connections": connections or Auth0Config.FALLBACK_VALUES["connections"]
         }
 
+    @function_cache(nocache={})
     def get_config_raw_data(self) -> dict:
         """
         Returns raw data (dictionary) from the Auth0 config URL.
-        This caches its data.
-        """
-        if not self._config_raw_data:
-            self._config_raw_data = self._get_config_raw_data_nocache() or {}
-        return self._config_raw_data
-
-    def _get_config_raw_data_nocache(self) -> dict:
-        """
-        Returns raw data (dictionary) from the Auth0 config URL.
-        This does NOT caches its data.
         """
         try:
             return requests.get(self.get_config_url()).json() or {}
@@ -166,9 +146,3 @@ class Auth0Config:
         headers = request.get("headers", {})
         scheme = headers.get("x-forwarded-proto", "http")
         return f"{scheme}://{domain}{context}callback/?react"
-
-    def cache_clear(self) -> None:
-        """
-        Clears out the caches, so next call to get_config_data() and get_config_raw_data() get fresh data.
-        """
-        self._config_data = self._config_raw_data = {}