+++ foursight_core/react/api/react_api.py
@@ -16,7 +16,7 @@ from dcicutils.env_utils import EnvUtils, get_foursight_bucket, get_foursight_bu
 from dcicutils.env_utils import get_portal_url as env_utils_get_portal_url
 from dcicutils.function_cache_decorator import function_cache, function_cache_info, function_cache_clear
 from dcicutils import ff_utils
-from dcicutils.misc_utils import ignored
+from dcicutils.misc_utils import get_error_message, ignored
 from dcicutils.obfuscation_utils import obfuscate_dict
 from dcicutils.redis_tools import RedisSessionToken, SESSION_TOKEN_COOKIE
 from dcicutils.ssl_certificate_utils import get_ssl_certificate_info
@@ -25,17 +25,16 @@ from .auth import AUTH_TOKEN_COOKIE
 from .auth import Auth
 from .aws_network import (
     aws_get_network, aws_get_security_groups,
-    aws_get_security_group_rules, aws_get_subnets, aws_get_vpcs, aws_network_cache_clear
+    aws_get_security_group_rules, aws_get_subnets, aws_get_vpcs
 )
 from .aws_s3 import AwsS3
 from .aws_stacks import (
     aws_get_stack, aws_get_stacks,
     aws_get_stack_outputs, aws_get_stack_parameters,
     aws_get_stack_resources, aws_get_stack_template,
-    aws_stacks_cache_clear
 )
 from .checks import Checks
-from .cognito import clear_cognito_cache, get_cognito_oauth_config, handle_cognito_oauth_callback
+from .cognito import get_cognito_oauth_config, handle_cognito_oauth_callback
 from .cookie_utils import create_delete_cookie_string, read_cookie, read_cookie_bool, read_cookie_int
 from .datetime_utils import convert_uptime_to_datetime, convert_utc_datetime_to_useastern_datetime_string
 from .encryption import Encryption
@@ -44,7 +43,6 @@ from .gac import Gac
 from .misc_utils import (
     get_base_url,
     is_running_locally,
-    memoize,
     sort_dictionary_by_case_insensitive_keys
 )
 from .portal_access_key_utils import get_portal_access_key_info
@@ -60,11 +58,6 @@ class ReactApi(ReactApiBase, ReactRoutes):
         super(ReactApi, self).__init__()
         self._react_ui = ReactUi(self)
         self._checks = Checks(app.core.check_handler.CHECK_SETUP, self._envs)
-        # self._cached_header = {}  # migrating to @function_cache
-        # self._cached_sqs_queue_url = None  # migrating to @function_cache
-        self._cached_accounts = None
-        self._cached_accounts_from_s3 = None
-        self._cached_elasticsearch_server_version = None
 
     @staticmethod
     def _get_stack_name() -> str:
@@ -73,12 +66,8 @@ class ReactApi(ReactApiBase, ReactRoutes):
         """
         return os.environ.get("STACK_NAME")
 
-    @function_cache(nocache_none=True)
+    @function_cache(nocache=None)
     def _get_sqs_queue_url(self):
-        # Leave just commented of for now short term as migrating to use of @function_cache.
-        # if not self._cached_sqs_queue_url:
-        #     self._cached_sqs_queue_url = app.core.sqs.get_sqs_queue().url
-        # return self._cached_sqs_queue_url
         return app.core.sqs.get_sqs_queue().url
 
     def _get_versions_object(self) -> dict:
@@ -100,18 +89,13 @@ class ReactApi(ReactApiBase, ReactRoutes):
                 "elasticsearch_dsl": get_package_version("elasticsearch-dsl")
             }
 
+    @function_cache(nokey=True, nocache=None)
     def _get_elasticsearch_server_version(self) -> Optional[str]:
-        if not self._cached_elasticsearch_server_version:
-            try:
-                connection = app.core.init_connection(self._envs.get_default_env())
-                es_info = connection.es_info()
-                elasticsearch_server_version = es_info["version"]["number"]
-                self._cached_elasticsearch_server_version = elasticsearch_server_version
-            except Exception:
-                pass
-        return self._cached_elasticsearch_server_version
+        connection = app.core.init_connection(self._envs.get_default_env())
+        es_info = connection.es_info()
+        return es_info.get("version", {}).get("number")
 
-    @memoize
+    @function_cache
     def _get_user_projects(self, env: str, raw: bool = False) -> list:
         """
         Returns the list of available user projects.
@@ -131,7 +115,7 @@ class ReactApi(ReactApiBase, ReactRoutes):
             ]
         return projects
 
-    @memoize
+    @function_cache
     def _get_user_institutions(self, env: str, raw: bool = False) -> list:
         """
         Returns the list of available user institutions.
@@ -156,7 +140,7 @@ class ReactApi(ReactApiBase, ReactRoutes):
             ]
         return institutions
 
-    @memoize
+    @function_cache
     def _get_user_roles(self, env: str) -> list:
         ignored(env)
         #
@@ -175,14 +159,14 @@ class ReactApi(ReactApiBase, ReactRoutes):
         ]
         return [{"id": role, "name": role, "title": role.replace("_", " ").title()} for role in roles]
 
-    @memoize
+    @function_cache
     def _get_user_schema(self, env: str) -> dict:
         portal_url = get_base_url(app.core.get_portal_url(env))
         user_schema_url = f"{portal_url}/profiles/User.json?format=json"
         user_schema = requests.get(user_schema_url).json()
         return user_schema
 
-    @memoize
+    @function_cache
     def _get_user_statuses(self, env: str) -> list:
         user_schema = self._get_user_schema(env)
         user_schema_properties = user_schema.get("properties") if user_schema else None
@@ -275,11 +259,6 @@ class ReactApi(ReactApiBase, ReactRoutes):
         # Note that this route is not protected but/and we return the results from authorize.
         auth = self._auth.authorize(request, env)
         data = self._reactapi_header_cache(request, env)
-        # Leave just commented of for now short term as migrating to use of @function_cache.
-        # data = self._cached_header.get(env)
-        # if not data:
-        #     data = self._reactapi_header_cache(request, env)
-        #     self._cached_header[env] = data
         data = copy.deepcopy(data)
         data["auth"] = auth
         # 2022-10-18
@@ -302,7 +281,7 @@ class ReactApi(ReactApiBase, ReactRoutes):
         if test_mode_certificate_simulate_error:
             data["portal"]["url"] = None
         # Note that we know that data["portal"]["url"] is explicitly set via _reactapi_header_cache.
-        if not data["portal"]["url"]:
+        if not data["portal"].get("url"):
             # Here we did not get a Portal URL from the to app.core.get_portal_url (via _reactapi_header_cache).
             # That call ends up ultimately calling the Portal health endpoint (via s3Utils.get_synthetic_env_config
             # via environment.get_environment_and_bucket_info). So there may have been a problem with the Portal,
@@ -448,12 +427,18 @@ class ReactApi(ReactApiBase, ReactRoutes):
             test_mode_access_key_expiration_warning_days=test_mode_access_key_expiration_warning_days)
         return self.create_success_response(response)
 
-    @memoize
+    def reactapi_elasticsearch(self) -> Response:
+        try:
+            return self.create_success_response(self._get_elasticsearch_server_version())
+        except Exception as e:
+            return self.create_error_response(get_error_message(e))
+
+    @function_cache
     def _get_env_and_bucket_info(self, env: str, stage_name: str) -> dict:
         return sort_dictionary_by_case_insensitive_keys(
             obfuscate_dict(app.core.environment.get_environment_and_bucket_info(env, stage_name)))
 
-    @memoize
+    @function_cache
     def _get_check_result_bucket_name(self, env: str) -> Optional[str]:
         envs = app.core.init_environments(env=env)
         if not envs:
@@ -1199,19 +1184,17 @@ class ReactApi(ReactApiBase, ReactRoutes):
 
     def _get_accounts_from_s3(self, request: dict) -> Optional[dict]:
         # Let's not cache this for now, maybe change mind later, thus the OR True clause below.
-        if True or not self._cached_accounts_from_s3:
-            s3_uri = self._get_accounts_file_from_s3()
-            if not s3_uri:
-                return self._get_accounts_only_for_current_account(request)
-            s3_uri = s3_uri.replace("s3://", "")
-            s3_uri_components = s3_uri.split("/")
-            if len(s3_uri_components) != 2:
-                return self._get_accounts_only_for_current_account(request)
-            bucket = s3_uri_components[0]
-            key = s3_uri_components[1]
-            accounts_json_content = AwsS3.get_bucket_key_contents(bucket, key)
-            self._cached_accounts_from_s3 = self._read_accounts_json(accounts_json_content)
-        return self._cached_accounts_from_s3
+        s3_uri = self._get_accounts_file_from_s3()
+        if not s3_uri:
+            return self._get_accounts_only_for_current_account(request)
+        s3_uri = s3_uri.replace("s3://", "")
+        s3_uri_components = s3_uri.split("/")
+        if len(s3_uri_components) != 2:
+            return self._get_accounts_only_for_current_account(request)
+        bucket = s3_uri_components[0]
+        key = s3_uri_components[1]
+        accounts_json_content = AwsS3.get_bucket_key_contents(bucket, key)
+        return self._read_accounts_json(accounts_json_content)
 
     def _get_accounts_only_for_current_account(self, request: dict) -> Optional[list]:
         aws_credentials = self._auth.get_aws_credentials(self._envs.get_default_env())
@@ -1243,19 +1226,18 @@ class ReactApi(ReactApiBase, ReactRoutes):
                     account["id"] = account_name
         return accounts_json
 
+    @function_cache(nocache=None)
     def _get_accounts(self) -> Optional[dict]:
         accounts_file = self._get_accounts_file()
         if not accounts_file:
             return None
-        if not self._cached_accounts:
-            try:
-                with io.open(self._get_accounts_file(), "r") as accounts_json_f:
-                    accounts_json_content_encrypted = accounts_json_f.read()
-                    accounts_json = self._read_accounts_json(accounts_json_content_encrypted)
-                    self._cached_accounts = accounts_json
-            except Exception:
-                return None
-        return self._cached_accounts
+        try:
+            with io.open(self._get_accounts_file(), "r") as accounts_json_f:
+                accounts_json_content_encrypted = accounts_json_f.read()
+                accounts_json = self._read_accounts_json(accounts_json_content_encrypted)
+                return accounts_json
+        except Exception:
+            return None
 
     def reactapi_accounts(self, request: dict, env: str, from_s3: bool = False) -> Response:
         ignored(env)
@@ -1576,33 +1558,6 @@ class ReactApi(ReactApiBase, ReactRoutes):
             function_cache_clear()
             cache_cleared.append("<all>")
         return self.create_success_response({"cache_cleared": cache_cleared})
-
-    def reactapi_clear_cache(self, request: dict) -> Response:
-        """
-        Called from react_routes for endpoint: GET /__clearcache___
-        Not yet implemented.
-        """
-        ignored(request)
-        app.core.cache_clear()
-        super().cache_clear()
-        self._checks.cache_clear()
-        self._react_ui.cache_clear()
-        Gac.cache_clear()
-        # self._cached_header = {}  # migrating to @function_cache
-        # self._cached_sqs_queue_url = None  # migrating to @function_cache
-        ReactApi._reactapi_header_cache.cache_clear()  # migrating to @function_cache
-        ReactApi._get_sqs_queue_url.cache_clear()  # migrating to @function_cache
-        self._cached_elasticsearch_server_version = None
-        self._cached_accounts = None
-        self._cached_accounts_from_s3 = None
-        self._get_env_and_bucket_info.cache_clear()
-        self._get_check_result_bucket_name.cache_clear()
-        self._get_user_projects.cache_clear()
-        self._get_user_institutions.cache_clear()
-        aws_network_cache_clear()
-        aws_stacks_cache_clear()
-        clear_cognito_cache()
-        return self.create_success_response({"status": "Caches cleared."})
 
     @staticmethod
     def reactapi_testsize(n: int) -> Response: