+++ react/src/pages/ApiCachePage.js
@@ -2,11 +2,12 @@ import { useState } from 'react';
 import Char from '../utils/Char';
 import { HorizontalLine } from '../Components';
 import Server from '../utils/Server';
+import Json from '../utils/Json';
+import Tooltip from '../components/Tooltip';
+import Yaml from '../utils/Yaml';
 import useFetch from '../hooks/Fetch';
 import useFetcher from '../hooks/Fetcher';
 import useFetchFunction from '../hooks/FetchFunction';
-import Json from '../utils/Json';
-import Yaml from '../utils/Yaml';
 
 const tdstyle = { fontSize: "11pt", verticalAlign: "top", paddingBottom: "2pt", paddingRight: "10pt", whiteSpace: "nowrap" };
 const tdstyler = { ...tdstyle, textAlign: "right" };
@@ -19,12 +20,20 @@ const ApiCache = (props) => {
         props.refresh();
     }
     return <>
+
         <tr>
-            <td style={tdstyle}>{cache.function}</td>
-            <td style={tdstyler}>{cache.hits}</td>
-            <td style={tdstyler}>{cache.misses}</td>
-            <td style={tdstyler}>{cache.size}</td>
-            <td style={tdstyle}>{cache.updated || <>&ndash;</>}</td>
+            <td style={tdstyle}>{cache.name}</td>
+            <td style={tdstyler}>{cache.hits > 0 ? cache.hits : <>&ndash;</>}</td>
+            <td style={tdstyler}>{cache.misses > 0 ? cache.misses : <>&ndash;</>}</td>
+            <td style={tdstyler}>{cache.size > 0 ? cache.size : <>&ndash;</>}</td>
+            <td style={tdstyle} id={`tooltip-updated-${cache.name}`}>
+                <Tooltip id={`tooltip-updated-${cache.name}`} position="bottom" text={"Last time function was actually called."} />
+                {cache.updated || <>&ndash;</>}
+            </td>
+            <td style={tdstyler} id={`tooltip-duration-${cache.name}`}>
+                <Tooltip id={`tooltip-duration-${cache.name}`} position="bottom" text={"Duration of last actual function call."} />
+                {cache.duration ? (cache.duration?.toFixed(1) + ' ms') : <>&ndash;</>}
+            </td>
             <td style={tdstyler}>
                 { cache.maxsize >= Number.MAX_SAFE_INTEGER ? <>
                     {Char.Infinity}
@@ -34,9 +43,10 @@ const ApiCache = (props) => {
             </td>
             <td style={tdstyler}>{cache.ttl || <>&ndash;</>}</td>
             <td style={tdstyler}>{cache.ttl_none || <>&ndash;</>}</td>
-            <td style={tdstyle}>{cache.nocache_none ? <>No</> : <>Yes</>}</td>
+            <td style={tdstyle}>{cache.nocache_none ? <>Yes</> : <>&ndash;</>}</td>
+            <td style={tdstyle}>{cache.nocache_other ? <>Yes</> : <>&ndash;</>}</td>
             <td style={tdstyle}>
-                <span onClick={() => clearCache(cache.function)} className="pointer">Clear</span>
+                <span onClick={() => clearCache(cache.name)} className="pointer">Clear</span>
             </td>
         </tr>
     </>
@@ -47,7 +57,7 @@ const ApiCachePage = (props) => {
     const cache = useFetch("//__functioncache__");
     const fetch = useFetchFunction();
     const [ showJson, setShowJson ] = useState(false);
-    const thstyle = { ...tdstyle, fontWeight: "bold", textDecoration: "underline" };
+    const thstyle = { ...tdstyle, verticalAlign: "bottom", fontWeight: "bold", textDecoration: "underline" };
 
     function refresh() {
         cache.refresh();
@@ -81,21 +91,23 @@ const ApiCachePage = (props) => {
             <table>
                 <thead>
                     <tr>
-                        <td style={thstyle}>Function</td>
+                        <td style={thstyle}>Name</td>
                         <td style={thstyle}>Hits</td>
                         <td style={thstyle}>Misses</td>
                         <td style={thstyle}>Size</td>
                         <td style={thstyle}>Updated</td>
+                        <td style={thstyle}>Duration</td>
                         <td style={thstyle}>Max Size</td>
                         <td style={thstyle}>TTL</td>
                         <td style={thstyle}>TTL None</td>
-                        <td style={thstyle}>Cache None</td>
+                        <td style={thstyle}>Null<br />NoCache</td>
+                        <td style={thstyle}>Other<br />NoCache</td>
                         <td style={thstyle}>Action</td>
                     </tr>
                 </thead>
                 <tbody>
                     { cache.map(item =>
-                        <ApiCache cache={item} key={item.function} refresh={refresh} />
+                        <ApiCache cache={item} key={item.name} refresh={refresh} />
                     )}
                 </tbody>
             </table>