+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Change Log
 
 * Support for catching SSL certificate errors for Portal.
 * Using new dcicutils.scripts.publish_to_pypi for publish.
+* Using new dcicutils.function_cache decorator for API function caching.
 
 4.0.0
 =====